Only test `after` date for "last changed" fields
If we only know the "last changed" date and not the actual date, we can't meaningfully determine whether the item (a site or page) contains a version that matches the `before` part of our date filter, so only check the `after` part of the filter in that case. This was causing us to miss data when looking for versions in non-current date ranges.