Don't calculate new file paths in formatters
Formatters used to independently calculate what path the outputted diffs/raw versions would be at, but that doesn't really work anymore since paths depend on the content of zip archives. Now we save the path that was used when writing output and the formatter simply uses that. This is arguably the way it should've been from the start.