Up capture vs. archive date window to 30 minutes
I figured 5 minutes would be plenty, but in practice I've now seen some higher than 25 minutes. So far, 30 minutes seems to work. Since we are always limiting to just versions of the same page, it might be safe to go to a few hours, but I'm a little worried about going that wide.