Add retryability to S3/Google uploads
Google seemed to be having frequent errors in production when uploading really large sets. We are mostly working with tiny files, so the overhead of Google's built-in resumable upload functionality is not worthwhile (it *really* slows things down), but this gives us an easy way to make a second or third attempt before finally giving up and crashing. That fixes the problem pretty consistently.