Get "pages" from all page type categories
Versionista has tabs that separate content into HTML, Images, and Scripts, which appears to happen based on the `Content-Type` header of the most recent version. It looks like all types used to be surfaced in the main list, but it's now only HTML (so we have some historic items from other, non-HTML tabs). It also seems that pages can move around between the tabs -- if an image suddenly returns a 404 or 500 error page with HTML in it, for example, it'll move from the Images tab to the HTML tab. If it later starts working again (not uncommon), then it'll move *back* to the Images tab.

This change also makes the default grabbing pages from all tabs, so we never lose versions because the moved between tabs.