Add option --group-by-site (fixes #3)
As you might guess, this causes the scraper to output a separate file for each site. This isn't the most beautiful code, but is a quick fix to be able to use this script for updated output for the tracking team, who is running into issues with the older script today.