Remove Versionista metadata from diff hash
Versionista adds some scripting, styling, and other extra bits at the top of a diff file. We now remove those bits from the part of the diff that gets hashed (and counted for length) so it's easier to see when there was no actual diff (now the length is 0 and text that is hashed is an empty string).