Show nothing in CSV for hashes of empty strings
The empty string's hash is actually a nice long string of random-looking text to most people, so to make it obvious in CSV output (which is already designed for human readers), show nothing when the hash is a hash of the empty string. This helps make it obvious that there is no actual diff.