HOTFIX: get raw content version by version
The zip archive functionality appears to have been removed (or removed specially for us!), so retrieve raw content version by version, similar to how we retrieve diffs. This was our very original approach, but had issues around file naming, exactitude of content, etc, so we abandoned it in favor of the zip archives, which were a better guarantee. This seems to be working alright now, though.