Log errors when Versions are missing dates.
It's OK for sites and pages to be missing `lastChanged` dates (it just makes things take much longer to run), but it's a problem if dates are missing from actual Version records. Now we look for this case and explicitly log an error (and fail the process), but still let the rest of the work finish in case only one page or version has issues.