Branch date checks based on object shape
We previously branched in `isInRequestedDateRange()` based on whether an object had a value for various types of date fields. Now that we can have `null` dates (to represent failed detection or parsing), we should branch based on the object *shape* (what properties it contains) instead, since we still want to catch those `null` values.