Regain some speed by checking `versions` column
The new changes to gracefully handle a broken `since` column in Versionista mean that we also check pages that Versionista has found but isn't even scraping! That wastes a lot of time and HTTP requests. However, the same listing view has a `versions` column that is a count of the number of versions that have been found for a page. If that column had a valid value, we can check whether it was greater 0 so we know whether it's really worth scraping the page.