Use a regex to remove Versionista HTML additions
It turns out our old exact-match check for these Versionista additions to the raw source of versions broke a while ago, so we have source now that does not match the original server response anymore. This fixes the issue when gathering new content by using a regular expression that is a little more flexible.

We still have to go clean up all our old records :\

Fixes #56.