Add "latest-to-base" links for our UI
The earliest (or "base") version of a page is found by querying for all the versions of each found page and selecting the earliest one.

This also adds a `--link-to-versionista` option. If used, links will point towards Versionista. If not used, links will point towards our UI (with the exception of the "view page" link, which is always Versionista for now -- see analyst meeting on 2017-01-01 for discussion). This allows us to easily reverse usage without needing to roll back the source.

Fixes #58.