HOTFIX: Stop pulling every page 3x
Also fix CSV formatter so it doesn't throw exceptions (can't believe we somehow had this left this way).

On pulling pages 3x: Versionista used to have a separate listing of "pages," "scripts," and "images," based on the content-type of a URL you were monitoring. Sometimes things would move between the lists (e.g. an error response would be a different content-type), so we had to check all three. At some point, this whole concept was removed from versionista and the URL for all three listing types returned the same listing! So we'd find a page 3 times and do all the analysis on it 3 times. YEESH.