HOTFIX: handle page links that don't redirect
Versionista's links to the actual, live page being monitored used to go through a redirect. We previously processed that URL and looked for the actual destination URL, which was after the querystring. Now they are simple direct links. For URLs without querystrings, this continued to function fine, but for those *with* querystrings, we wound up cutting off everything before the querysting of the destination!

For example, this:

    http://research.noaa.gov/SecureContentAdd/tabid/405/Default.aspx?returnurl=%2FHome%2FPrivacyPolicy.aspx

became:

    returnurl=%2FHome%2FPrivacyPolicy.aspx

an obviously incorrect and invalid URL.

The new strategy checks to see if the literal URL is within Versionista, in which case we assume it functions like a redirect. Otherwise we just take the URL as-is.