HOTFIX: failure on table redesign in Versionista
Versionista redesigned the table layout on the page history view (`https://versionista.com/{site}/{page}/`) so that every other row is a hidden form pertaining to the version in the row above it. We were tripping over these non-version rows. This change uses a class that was added to the version rows in order to identify and scan *just* them. It also gets a little stricter on handling the situation so similar issues are more obvious in the future:
- If we don't find any version rows, we print a warning.
- If we don't find a node representing the date in a row, we throw a more descriptive error (previously we just got "can't read property `textContent` of `undefined`").