HOTFIX: handle change to date column on site view
The columns in the table listing pages for a given site were re-arranged, resulting in us looking for a "last change" timestamp in the column that formats changes as "x minutes ago". This fixes which column we are looking at, but also adds a bunch more sanity checking to let us know if something similar occurs again.