Auto-retry requests for more types of errors
In production, we were seeing `EAI_AGAIN` (temporary failure) errors that could potentially have been resolved by retrying the request. Instead of *just* adding that, I took a look at the error codes LibUV might spit out and attempted to pick a better list of ones that might make sense to retry on. It's likely not perfect, but should be a good reliability improvement.

Fixes #129.