HOTFIX: handle line breaks in page titles
It turns out some of our page titles have line breaks in them, which should not be the case. It turns out to be a real situation we have, though, so this handles it by cleaning up the title strings to be treated like HTML would treat titles.

Additionally, it's always possible that the data we put in a CSV could have a line break in a cell (or a quote!), so actually make sure to properly escape both of those cases.