Enforce `--chunk-delay` getting earliest version
We already obeyed `--chunk-delay` for any request that iterate over multiple chunks (hence the name), but if it's meant to impose reasonable rate limiting, we should have been doing it when fetching the first version of every found page, too. (In fact, not doing this caused this script to fail a lot this week.)

A better fix is upcoming in edgi-govdata-archiving/web-monitoring-db#438 (which we'll have to write code to utilize here) -- that will let us get the earliest version without making an extra request for every page. This just tides us over for now.