Utilize new `?include_earliest` query param
We spend a stupid amount of time and effort finding the earliest version of each page when generating analyst sheets. Now -db has a PR for adding an `?include_earliest` option, which will do all that work for us! This makes use of it, simplifying and speeding things up nicely :)

See also edgi-govdata-archiving/web-monitoring-db#438