Use Versionista CSV data to add more metainfo
Versionista now has a CSV you can download for each page that lists useful extra metadata, like content type and response code (and a few others)! This is SUPER useful, since we can't reliably get some of this data from the rest of the page. Unfortunately it doesn't include *everything* relevant, so this downloads the CSV alongside the page, then combines the data we get from each.

I'm not sure when this feature got added -- there used to be a CSV for the whole account (and it's still there), but it has even less info and does not fill us in with what we need, so we skip over it. It's a different story with the page-level one, though.