Add script to load CSV metadata for old versions
This looks through every page in a list of pages, scrapes its data from Versionista (including the new CSV data), and posts a merge update to web-monitoring-db with just the relevant stuff from the CSV that we weren't able to load in the past. This is pretty much the best kind of backfill we can do.