HOTFIX: handle missing "last updated since" field
Some sites just started not having a "last updated since" field. It *seems* like this correlates to not having any updates in the past year or so (though I could be totally wrong there...), and if it's the case, there is a `class="anev"` element (maybe the class is for "never updated?") instead. This hotfix changes things to assume that means the last update was probably a year ago. I think this is reasonable, though the more conservative case would be to mark it as having potentially just been updated so the site always gets checked.