Fail hard if sites or pages listing have no tables
Judging by what's happening on the page details view where the list of versions is no longer in the requested HTML page (instead it comes via a secondary XHR), a similar change is likely in the works for the list of sites and the list of pages. This adds code to fail hard if those URLs stop containing anything that looks like content so we know right away when we have to change how they work.