HOTFIX: versions are no longer in the HTML
It looks like Versionista has changed up the page details view where versions are listed. Instead of delivering them as content in the HTML page, there is now an API (!) that gets called to load the list of versions via XHR after the page loads. Unfortunately, the API doesn't appear to extend beyond the list of versions and root page of the API says:

> Our API is still very much a work in progress and is in beta, so it is not feature-complete and is subject to revision.

So I've treated it similar to scraping and added a lot of extra error checking around it. Overall this is nice, though! We get structured data (even if we are guessing a bit at the various fields, some of which I couldn't identify) that includes almost everything we got from the HTML + CSV combined (now in only one request!). The only thing missing is the load time, which the CSV has. We don't make active use of that info and we haven't been gathering it from the CSVs for that long, though, so for now I'm just dropping it. We can always go back to making the CSV request and combining the two results if we think it's useful, though.

Fixes #195.