Handle the fact that the page details view no longer has content (#196)
The page view in Versionista (e.g. https://versionista.com/73900/6185299/) no longer includes the list of versions. Instead, versions are now loaded via XHR in JavaScript after the page loads. This PR is the quickest reasonable fix to get things working again; it probably needs cleanup later.

Versionista now loads the list of pages from a JSON API located at `https://versionista.com/api/versions/{site_id}/{page_id}/`. I poked around and didn’t see any obvious corresponding endpoints for sites or pages, but I’d bet those are probably coming soon.

To that end, this adds code to detect and throw exceptions if we see similar issues on the list of sites or the list of pages in a site. That’ll trigger alerts about failed scrapes.

This also does some potentially excessive error checking around the expected schema of the API we’re calling (since the API describes itself as not yet being final, we should certainly be careful about potential changes). On the upside, the API includes almost everything we used to get from the HTML + CSV combined (now in only one request!). The only thing missing is the load time, which the CSV has. We don't make active use of that info and we haven't been gathering it from the CSVs for that long, though, so for now I'm just dropping it. We can always go back to making the CSV request and combining the two results if we think it's useful, though.

Fixes #195.