Show links/stats for whole period in task Sheets
In the weekly tasking sheets we generate for analysts, we show a diff link that compares the latest two versions and show diff hashes/lengths/priority for the change between those same last two versions. However, a page may have gone through several changes in the queried time period (usually a week), and that means the data we're showing is a much less complete description of what an analyst should be considering than it could be.

I'd been thinking of doing something like this for a while, but put it off because I thought it would be hard, and there were a lot of other tasks on my plate. During a discussion with analysts this week, though, I realized that we could leverage the change objects we now have to do this without rethinking how we query things! If a page couldn't be auto-analyzed (e.g. it's a Word doc, image, PDF, etc.) we'll still have the same old behavior as before, BUT if we have change information, we can do better:

- The diff links for "last two" to "this period" point to the diff between the last version BEFORE the query timeframe and the latest version in the query timeframe. That means the diff it links two is the diff across all the changes in the timeframe.

- The stats (diff hash, diff length, priority) are now summaries that account for all the changes in the timeframe. It's impractical to do a full analysis of lots of pages for an arbitrary timeframe (and this script is arbitrary; though it's generally set for a week, we occasionally make changes), so we attempt to summarize each of the version-to-version changes involved in the timeframe in a reasonable way.

Fixes #208.