Handle diffs properly when scripts are enabled
Versionista now has an "enable scripts" option, which is a querystring in the actual URL. We were accidentally appending the diff type to the end of the querystring instead of the end of the path (which is *before* the querystring), causing malformed URLs. This uses the right values so everything works.

Fixes #222.