If tag names match exactly, include full tag name
We only use the end of the name if the match was a prefix. In exact match cases, it would be an empty string, and that would just be confusing in the output. This is mainly because we are now use non-"categorical" tags for sorting. In particular, we now have blogs and news feeds tagged with "news", which is different from things like "2l-domain:<whatever>". In the old code, a page tagged with "news" would have "" (empty string) for the news tag's part of the group name in the output. Now it actually shows "news".