If tag names match exactly, use full tag name file names (#226)
When matching a page’s tags during DB queries (using the `--group-by` option), we treat the option as a prefix, and only include everything *after* the prefix in our output. For example, if using `--group-by '2l-domain:'`, and matching a page tagged `2l-domain:epa.gov`, we’d put the page in the `epa.gov` group (everything after the matched prefix).

Now, however, we are starting to use tags that aren’t “categorical” the way `2l-domain:<xyz>` is. For example, we want to group on the `news` tag, which designates whether a page is a blog/newsfeed/etc. The tag is merely present or not, rather than designating a standard prefix with different suffix values.

To support that, this change uses the full tag name (instead of the suffix) in situations where the match is an exact one, rather than just a prefix. So, when matching `2l-domain:`, we still use the group `epa.gov`, but when matching `news`, we use `news` as the group, because it is the complete tag name.