Skip intermittent errors when scoring pages (#230)
When calculating the overall priority and change size for a page over the course of a given time period, we consider every version we recorded in that period. However, if there was an intermittent error (e.g. one scrape in the middle of the time period got a 500 error that was later resolved), we often wind up with inflated priorities and change sizes because it looks like the whole page changed multiple times. As an imperfect solution, we now skip over versions that were errors or recoveries from errors (unless the ending state was an error, in which case we *do* want to count them).

This is a little hacky, and it's compounding the already hacky way we calculate overall scores for the week. At some point we probably need to find an effective way to make sure we calculate *actual* change scores for the given period instead of just adding up all the one-by-one scores.