Use a cache for API requests so we can resume a broken analyst s… (#231)
Getting a full run of the analyst’s task sheets is really becoming problematic because the database is often dying in the middle. The real solution for that is [edgi-govdata-archiving/web-monitoring-db#548](https://github.com/edgi-govdata-archiving/web-monitoring-db/pull/548), but in the mean time, as an emergency fix, this adds a caching mechanism for all our API requests. That way, we can resume a broken job from halfway, and not put the same load right back on the database again.

The basic idea here is to wrap all the HTTP requests in `withCache`, which loads the cache and returns a cached result if present or, if not, runs the underlying function (the HTTP request) and caches the result. Every five seconds, it pesists the cache to disk.