Update selector for a site's last change time (#236)
There used to be a unique ID we could use to identify the cell holding the time since the last change on a site, but that appears to be gone. There's really no identifying ID, class, or other attribute anymore, so this now just uses the next cell after the "new pages found" cell.