Test live site in CI (#238)
Since Versionista updates its UI every so often, it might help to run a scrape nightly and make sure there are no errors. That’s not perfect (there are plenty of silent errors when Versionista changes, too), but it’s a lot better than nothing!

I’ve made a new free account just for testing that only has a few pages, but which will update regularly, so this should always grab some data.