HOTFIX: the pages of a site now come from an API
Ever since the data in the page details view got replaced by an API, we've been waiting for the same to happen to the site details view. Now it has! This logic mostly mirrors the API logic we wrote for the page details view.

I've also expanded the type checking a bit to:
- Handle optional properties (we just didn't check their types before).
- Handle arrays (we don't check what types they are arrays *of*, though).