Remove `Versionista.getVersionArchiveEntries()` (#272)
Remove the `getVersionArchiveEntries(pageUrl)` method from the `Versionista` class (and the `unzip-stream` package it used). We haven't used this method of downloading raw content in at least 1.5 years; it had major drawbacks with respect to `getVersionRawContent()`. (Since we haven't used it in so long and it doesn't have tests, I'm also not even sure it still works!)

NOTE: `reload-versionista-archives` script has been removed -- it depended on `getVersionArchiveEntries()`, but was already deprecated.

Fixes #90.