HOTFIX: Use CSV for sites list instead of scraping (#383)
We used to scrape the content of the main page that listed all the sites to get a list of sites. That page is now empty and populated entirely by JS, so we now get this information by downloading a CSV of all URLs in the account and boiling down a list of sites from that.